+++ simplejrpc/schemas.py
@@ -3,8 +3,6 @@
 from wtforms import Form, validators
 from wtforms.fields import simple
 
-import simplejrpc.config as setting
-from simplejrpc.i18n import GI18n
 from simplejrpc.i18n import T as i18n
 from simplejrpc.interfaces import BaseValidator
 
@@ -12,12 +10,13 @@ from simplejrpc.interfaces import BaseValidator
 class StringLangValidator(BaseValidator):
     """ """
 
-    def __init__(self, lang="en"):
+    def __init__(self, lang="en", err_message=None):
         self.lang = lang
+        super().__init__(err_message)
 
     def validator(self, form, field):
         lang = field.data or self.lang
-        GI18n(setting.PROJECT_I18n_PATH, lang)
+        i18n.set_lang(lang)
         return lang
 
 
@@ -27,27 +26,32 @@ class StrRangeValidator(BaseValidator):
     allows = []
     err_message = ""
 
-    def __init__(self, allows, message=None):
+    def __init__(self, allows, err_message=None):
         """ """
         self.allows = allows
-        self.err_message = message or self.err_message
+        super().__init__(err_message)
 
     def validator(self, form, field):
         if field.data not in self.allows:
-            message = i18n.translate(self.err_message) if self.err_message else f"expected value {self.allows}"
+            message = (
+                i18n.translate(self.err_message)
+                if self.err_message
+                else f"expected value {self.allows}"
+            )
             raise validators.ValidationError(message)
 
 
-class IntLimitValidator:
+class IntLimitValidator(BaseValidator):
     """ """
 
     min: int
     max: int
     err_message = ""
 
-    def __init__(self, min=1, max=1000):
+    def __init__(self, min=1, max=1000, err_message=None):
         self.max = max
         self.min = min
+        super().__init__(err_message)
 
     def validator(self, form, field):
         """ """