+++ CHANGELOG.md
@@ -4,6 +4,19 @@ Changelog
 Next Version
 ------------
 
+2.0.4 (2025-05-13)
+------------------
+* Fix: Write response format
+
+2.0.2 (2025-05-13)
+------------------
+* Fix: Read config content error
+
+2.0.1 (2025-05-12)
+------------------
+* Add: GmRequest
+* Doc: Update doc import method
+
 2.0.0 (2025-05-12)
 ------------------
 * Add: project detail